Initial commit
This commit officially publishes the ibp-ng source code to the community.
At present, ibp-ng is an experimental software tool, but changes will be
pushed over time to bring it closer to completion.

It is expected that one or more parallel branches will evolve within this
repository to play with more 'bleeding-edge' mathematics and algorithms
related to iBP and the MDGP.