Added forcefield back-calculation
This commit includes code that back-calculates force field parameters for
use during energy pruning.

Currently, the back-calculation proceeds by interpreting the interval bounds
(or the exact value plus or minus the DDF tolerance) as the 2.5% and 97.5%
quantiles of a probability distribution, whose shape depends on the data:

 * Bonds: normal.
 * Non-bonded distances: log-normal.
 * Angles: von Mises.
 * Dihedrals: von Mises.
 * Impropers: von Mises.

Using quantiles is a bit odd for low-precision von Mises distributions,
as it will consistently over-estimate the precision of the underlying
distribution (it assumes approximate Gaussianity). The simplest solution
is to adjust the intervals in the ibp-protein.par file to force the
precision down near zero for un-informative torsions. However, this
will require careful inspection of the code to ensure that invalid
intervals don't make their way into the graph computations.

The next stages of this work are:

 * Implementation of an energetic pruning device
 * Pre-computation of discretization parameters, for:
    - Biasing the search using locally optimal energetics
    - Intelligently computing the set of d,theta,omega branch values