Initial energetic pruning features
This commit introduces the first functioning implementation of energetic
pruning in ibp-ng. At the moment, all terms except VDW and electrostatics
are computed based on the interval restraints coming from the user. The
force field parameters (force constant and mean distance/angle/etc) are
back-calculated from intervals by treating them as 95% confidence
intervals of probability distributions.

The selection of discretization point order (i.e. which branches to
traverse first) is also changed in this commit. Now innermost branches
are taken first in an attempt to enumerate better solutions earlier in
the tree search.

Finally, this commit introduces modified repetition orders for the
backbone-only residues BB2 and BBI. Now, iBP branches directly on the
phi and psi angles of the peptide backbone. This is an attempt to
alleviate problems that arose when precise dihedral restraints were
introduced. Using the old re-orders, phi/psi were highly nonuniformly
sampled, leading to high pruning rates, even when the discretization
factor was set high.