Modified output atom ordering
This commit introduces small changes to the graph and enumerator data
structures in order to write more "well-formed" DCD and PDB files.
Previously, the atoms were written in a sequence defined by the
repetition order, but this caused problems for [poorly written]
software like STRIDE that expected a certain consistency. The
new output sequence is based on the ordering of atoms within
the topology files.