Added further energetic pruning features
This commit includes a van der Waals contact term into the
energetic pruning device, completing the set of non-negative
terms that are to be included for energetic pruning. It also
modifies the bounds-checking performed on dihedral restraints,
allowing "out-of-bounds" restraints to be specified in order
to soften the corresponding energetic terms of those dihedrals,
while ensuring the dihedrals are bounded before inclusion into
the final graph data structure.