Modified energetic pruning for reporting
This commit modifies the energetic pruning check to perform
short-circuit logical tests of the energy at the current node.
Therefore, it is not necessary to complete the summation of all
energy terms at a node if the sum has already exceeded the known
threshold.

This commit also adds tracking information to the energetic
pruning data structures.