Implemented more intelligent branch control
This commit adds a bit more intelligence to the way iBP branches.
First, all graph edges are now tagged with semantic information,
and given a pointer back to the source value that generated them.

In other words, iBP now knows whether a graph edge came from a
distance, an angle, or a dihedral, and can decide how to branch
accordingly during traversal.

Paired with the new re-orders, this is a powerful way to reduce
the search space when a substantial amount of backbone dihedrals
are known with any degree of certainty.