Corrected mis-calculation of sub-tree widths
This commit introduces a proper (approximate) computation of the
sub-tree width during traversal, used for timing reports. This
method uses 'long double' types to achieve a general-purpose
usefulness, up to trees of size 10^4000 or so. :)

The calculations are still incorrect for multi-threaded
traversals, but this can either be fixed or not. The
usefulness of multi-threaded traversal is questionable.