Ironed out a bug with torsions, more tests
This commit finishes the 'data/orders' directory, which shows
how the new repetition order uniformly and adaptively samples
Ramachandran space, when dihedral constraints are available.

A bugfix is also included for the TAF (dihedral and improper)
pruning device, which was pruning backbone omega angles that
were -180 degrees, just because the feasible interval is
+180 +/- [tol] degrees. The TAF device now checks the
following intervals:

 [L - tol, U + tol] + {-2, 0, +2} * pi

where [L, U] is the original interval.