Added tests of loop-induced complexity
This commit adds the data/loops directory, which
contains scripts for testing the relationship
between loop flexibility, distance restraints,
and runtime to the first solution.

The tests are based on the result produced by
data/talos, so variations in local geometry are
non-existent, and solutions are guaranteed to
exist for each problem instance.