Made graph completion optional
This commit adds a command-line flag for enabling the use of graph
completion. By default, iBP-ng will not complete the graph, using
only the information directly available from the instance.

If the --complete flag is passed to iBP-ng, all missing edges in
the graph will be computed, using VDW for lower bounds and DSP
for upper bounds.

As it has been decided that path pruning using this complete graph
has no effect (as it doesn't use the partial realization to build
an auxiliary graph), this feature will be disabled by default,
and potentially removed in later versions.

The path pruning device will be replaced with a 'future' pruning
device, that will create a callback for each triple of atoms
(i,j,k : i < j < k in the order), where distance information
exists for (j,k) and (i,k).