Added a 'future' pruning device
This commit adds the 'future' pruning device, which checks the
triangle inequality for triangles formed by vertices i < j < k,
with x(i) and x(j) embedded, and x(k) in the future.

Further work is required to validate this pruning device, as
it seems that there are cases, e.g. when (i,j,k) are a clique
in the graph with exact weights, where we should avoid registering
a pruning callback.