Added graph edge refinement
This commit adds support for refining the edges of a graph using
information from triangle inequalities. The idea is identical to
future pruning, but instead of pruning, corrects the graph to
improve the efficiency of discretization.

It is highly likely that the path pruning device will be phased
out in future commits, as it no longer serves a purpose.