Ripped src/resid.* from the source tree [BORKED]
This commit breaks everything. The mapping between single-letter
and three-letter amino acid codes was moved into the FASTA parsing
source file. Basically, this is the beginning of a push to purely
rely on the topology specification for building peptides.

As it stands, the functions in src/resid.c were buried pretty
deeply into the code base, so linking fails with a bunch of
errors about undefined symbols, obviously.

The next commit will correct this issue.