Removed hard-coded residue types
Shockingly enough, until this commit, iBP-ng only supported the
twenty canonical amino acids, as a ton of the code ran through a
hard-coded table of single-letter and three-letter codes in order
to 'validate' residue types.

That's fixed now. Valid residue types are now the set of residues
defined in the topology file.